backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
groq
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
